move RA and RAC into a libs folder to resolve local linting and build issues
