+++ docs/.storybook/preview.tsx
@@ -7,7 +7,7 @@ import React, { useEffect } from 'react'
 import { decorators as bgDecorators } from '@storybook/addon-backgrounds/preview'
 import { type Decorator, type Preview } from '@storybook/react'
 import { KaizenProvider } from '~components/KaizenProvider'
-import { I18nProvider } from '~components/__react-aria-components__'
+import { I18nProvider } from '~components/__libs__/react-aria-components'
 import { ReversedColors } from '~components/utils'
 import { DefaultDocsContainer } from '../components/DocsContainer'
 import { backgrounds } from '../utils/backgrounds'
