+++ packages/components/libs/react-aria/package.json
@@ -1,0 +1,5 @@
+{
+  "main": "../../dist/cjs/reactAriaKZ.cjs",
+  "module": "../../dist/esm/reactAriaKZ.mjs",
+  "types": "../../dist/types/__react-aria__/index.d.ts"
+}
