--- packages/components/react-aria-components/package.json
@@ -1,5 +1,0 @@
-{
-  "main": "../dist/cjs/reactAriaComponents.cjs",
-  "module": "../dist/esm/reactAriaComponents.mjs",
-  "types": "../dist/types/__react-aria-components__/index.d.ts"
-}
