+++ packages/components/rollup.config.mjs
@@ -5,8 +5,8 @@ export default rollupConfig({
     index: './src/index.ts',
     next: './src/__next__/index.ts',
     alpha: './src/__alpha__/index.ts',
-    reactAria: './src/__react-aria__/index.ts',
-    reactAriaComponents: './src/__react-aria-components__/index.ts',
+    reactAriaKZ: './src/__libs__/react-aria/index.ts',
+    reactAriaComponentsKZ: './src/__libs__/react-aria-components/index.ts',
   },
   plugins: pluginsSharedUi,
 })
