+++ pnpm-workspace.yaml
@@ -2,8 +2,6 @@ packages:
   - 'packages/**'
   - 'docs'
   - '!packages/components/next'
-  - '!packages/design-tokens/js'
-  - '!packages/react-aria'
-  - '!packages/react-aria-components'
-  - '!packages/design-tokens/js'
   - '!packages/components/alpha'
+  - '!packages/components/libs'
+  - '!packages/design-tokens/js'
